# gh #427 Update test_l3_dsHdmiIn.c