Get the supported game feature and call the register API for VRR and ALLM only if its supported by platform