@thanushreevani 2025-05-15
**Before Fix:**
`Enter number of test to select (1-18) : 1

2025-05-15-17:12:49, LOG   , ut_console.c,   633 :

2025-05-15-17:12:49, LOG   , ut_console.c,   636 : Running Suite : [L3 dsHdmiIn]
2025-05-15-17:12:49, LOG   , ut_console.c,   639 :      Running Test : 'Initialize HdmiIn'
2025-05-15-17:12:49, INFO  , test_l3_dsHdmiIn.c,   450 : In test_l3_HdmiIn_initialize [03001]
2025-05-15-17:12:49, INFO  , test_l3_dsHdmiIn.c,   454 : Select Device type
         Sink: 0
         Source: 1
Select Device:0

2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   467 : Calling dsHdmiInInit()
250515-17:12:50:934726 [dsHal]dsHdmiInInit:1539  Initiializing tvclient observer
2025-05-15 17:12:50:934 22941 22941 D TV [TvClientWrapper]: GetInstance: start GetInstance.
2025-05-15 17:12:50:935 22941 22941 D TV [TvClient]: TvClient.
2025-05-15 17:12:50:936 22941 22941 D TV [TvClient]: Connected to tvservice.
2025-05-15 17:12:50:937 22941 22941 D TV [TvClient]: setTvClientObserver
2025-05-15 17:12:50:937 22941 22941 D TV [TvClient]: setTvClientObserver: observer is 0xa01d38.

250515-17:12:50:937234 [dsHal]checkHdmiInPortStatus:1466  ---> hdmi_status : 4

250515-17:12:50:937278 [dsHal]checkHdmiInPortStatus:1493  ---> dsHDMI_IN_PORT_0 connected

250515-17:12:50:937323 [dsHal]dsHdmiInInit:1564  <--- error=0

2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   469 : Result dsHdmiInInit() dsError_t:[dsERR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   476 : Calling dsHdmiInRegisterConnectCB(IN:CBFunc:[0x8BAD5])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   478 : Result dsHdmiInRegisterConnectCB(IN:CBFunc:[0x8BAD5]) dsError_t:[dsERR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   484 : Calling dsHdmiInRegisterSignalChangeCB(IN:CBFun:[0x8BA89])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   486 : Result dsHdmiInRegisterSignalChangeCB(IN:CBFun:[0x8BA89]) dsError_t:[dsERR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   491 : Calling dsHdmiInRegisterStatusChangeCB(IN:CBFun:[0x8B9E1])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   493 : Result dsHdmiInRegisterStatusChangeCB(IN:CBFun:[0x8B9E1]) dsError_t:[dsERR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   498 : Calling dsHdmiInRegisterVideoModeUpdateCB(IN:CBFun:[0x8B91D])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   500 : Result dsHdmiInRegisterVideoModeUpdateCB(IN:CBFun:[0x8B91D]) dsError_t:[dsERR_NO
NE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   505 : Calling dsHdmiInRegisterAllmChangeCB(IN:CBFun:[0x8B8D1])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   507 : Result dsHdmiInRegisterAllmChangeCB(IN:CBFun:[0x8B8D1]) dsError_t:[dsERR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   512 : Calling dsHdmiInRegisterAVLatencyChangeCB(IN:CBFun:[0x8B809])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   514 : Result dsHdmiInRegisterAVLatencyChangeCB(IN:CBFun:[0x8B809]) dsError_t:[dsERR_NO
NE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   519 : Calling dsHdmiInRegisterAviContentTypeChangeCB(IN:CBFun:[0x8B885])
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   521 : Result dsHdmiInRegisterAviContentTypeChangeCB(IN:CBFun:[0x8B885]) dsError_t:[dsE
RR_NONE]
2025-05-15-17:12:50, INFO  , test_l3_dsHdmiIn.c,   526 : Calling dsHdmiInRegisterVRRChangeCB(IN:CBFun:[0x8B839])./hal_test_dshal: symbol
lookup error: ./hal_test_dshal: undefined symbol: dsHdmiInRegisterVRRChangeCB
`

**After fix:**
`Enter number of test to select (1-18) : 1

2025-05-15-17:35:09, LOG   , ut_console.c,   633 :

2025-05-15-17:35:09, LOG   , ut_console.c,   636 : Running Suite : [L3 dsHdmiIn]
2025-05-15-17:35:09, LOG   , ut_console.c,   639 :      Running Test : 'Initialize HdmiIn'
2025-05-15-17:35:09, INFO  , test_l3_dsHdmiIn.c,   450 : In test_l3_HdmiIn_initialize [03001]
2025-05-15-17:35:09, INFO  , test_l3_dsHdmiIn.c,   454 : Select Device type
         Sink: 0
         Source: 1
Select Device:0

2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   467 : Calling dsHdmiInInit()
250515-17:35:11:666703 [dsHal]dsHdmiInInit:1539  Initiializing tvclient observer
2025-05-15 17:35:11:666 25346 25346 D TV [TvClientWrapper]: GetInstance: start GetInstance.
2025-05-15 17:35:11:666 25346 25346 D TV [TvClient]: TvClient.
2025-05-15 17:35:11:668 25346 25346 D TV [TvClient]: Connected to tvservice.
2025-05-15 17:35:11:668 25346 25346 D TV [TvClient]: setTvClientObserver
2025-05-15 17:35:11:668 25346 25346 D TV [TvClient]: setTvClientObserver: observer is 0x24ed38.

250515-17:35:11:669027 [dsHal]checkHdmiInPortStatus:1466  ---> hdmi_status : 4

250515-17:35:11:669059 [dsHal]checkHdmiInPortStatus:1493  ---> dsHDMI_IN_PORT_0 connected

250515-17:35:11:669091 [dsHal]dsHdmiInInit:1564  <--- error=0

2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   469 : Result dsHdmiInInit() dsError_t:[dsERR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   476 : Calling dsHdmiInRegisterConnectCB(IN:CBFunc:[0x66D15])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   478 : Result dsHdmiInRegisterConnectCB(IN:CBFunc:[0x66D15]) dsError_t:[dsERR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   484 : Calling dsHdmiInRegisterSignalChangeCB(IN:CBFun:[0x66CD5])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   486 : Result dsHdmiInRegisterSignalChangeCB(IN:CBFun:[0x66CD5]) dsError_t:[dsERR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   491 : Calling dsHdmiInRegisterStatusChangeCB(IN:CBFun:[0x66C59])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   493 : Result dsHdmiInRegisterStatusChangeCB(IN:CBFun:[0x66C59]) dsError_t:[dsERR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   498 : Calling dsHdmiInRegisterVideoModeUpdateCB(IN:CBFun:[0x66BA9])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   500 : Result dsHdmiInRegisterVideoModeUpdateCB(IN:CBFun:[0x66BA9]) dsError_t:[dsERR_NO
NE]DS HAL : Using /etc/hdmi_game_features.ini

250515-17:35:11:670211 [dsHal]DS-HAL(Debug): dsGetSupportedGameFeaturesList: Total Features : 1

250515-17:35:11:670251 [dsHal]DS-HAL: dsGetSupportedGameFeaturesList: Total Features: 1

250515-17:35:11:670278 [dsHal]DS_HAL: dsGetSupportedGameFeaturesList: Features List: ALLM

2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   509 : Calling dsHdmiInRegisterAllmChangeCB(IN:CBFun:[0x66B69])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   511 : Result dsHdmiInRegisterAllmChangeCB(IN:CBFun:[0x66B69]) dsError_t:[dsERR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   517 : Calling dsHdmiInRegisterAVLatencyChangeCB(IN:CBFun:[0x66AC1])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   519 : Result dsHdmiInRegisterAVLatencyChangeCB(IN:CBFun:[0x66AC1]) dsError_t:[dsERR_NO
NE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   524 : Calling dsHdmiInRegisterAviContentTypeChangeCB(IN:CBFun:[0x66B29])
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   526 : Result dsHdmiInRegisterAviContentTypeChangeCB(IN:CBFun:[0x66B29]) dsError_t:[dsE
RR_NONE]
2025-05-15-17:35:11, INFO  , test_l3_dsHdmiIn.c,   540 : Out test_l3_HdmiIn_initialize
2025-05-15-17:35:11, LOG   , ut_console.c,   661 :      Test Complete : 'Initialize HdmiIn'

Run Summary:    Type  Total    Ran Passed Failed Inactive
              suites      9      0    n/a      0        0
               tests    103      1      1      0        0
             asserts      0      0      0      0      n/a

Elapsed time =    0.005 seconds

***************** CUNIT CONSOLE - SUITE MENU ***************************
(R)un (S)elect (L)ist (A)ctivate (F)ailures (U)p (O)ptions (H)elp (Q)uit
`