@JayaPrakash-Madiraju 2025-05-15
Verified the changes in both ALLM and VRR supported and not supported platforms, there is no crash seen at Init.

[VRR_ALLM_API_remove.txt](https://github.com/user-attachments/files/20234117/VRR_ALLM_API_remove.txt)